+++ chaosaws/elbv2/actions.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 import random
-from collections import Counter
-from copy import deepcopy
-from typing import Any, Dict, List
+from typing import Dict, List
 
 import boto3
+from botocore.exceptions import ClientError
+
 from logzero import logger
 
 from chaosaws import aws_client
@@ -12,7 +12,7 @@ from chaosaws.types import AWSResponse
 from chaoslib.exceptions import FailedActivity
 from chaoslib.types import Configuration, Secrets
 
-__all__ = ["deregister_target"]
+__all__ = ["deregister_target", "set_security_groups", "set_subnets"]
 
 
 def deregister_target(tg_name: str,
@@ -32,9 +32,157 @@ def deregister_target(tg_name: str,
     return client.deregister_targets(TargetGroupArn=tg_arn[tg_name],
                                      Targets=[{'Id': random_target}])
 
+
+def set_security_groups(load_balancer_names: List[str] = None,
+                        security_group_ids: List[str] = None,
+                        configuration: Configuration = None,
+                        secrets: Secrets = None) -> List[AWSResponse]:
+    """
+    Changes the security groups for the specified load balancer(s).
+    This action will replace the existing security groups on an application
+    load balancer with the specified security groups.
+
+    Parameters:
+        - load_balancer_names: a list of load balancer names
+        - security_group_ids: a list of security group ids
+
+    returns:
+        [
+            {
+                'LoadBalancerArn': 'string',
+                'SecurityGroupIds': ['sg-0000000', 'sg-0000001']
+            },
+            ...
+        ]
+    """
+    if not load_balancer_names:
+        raise FailedActivity('You must specify at least 1 load balancer.')
+
+    if not security_group_ids:
+        raise FailedActivity('You must specify at least 1 security group id')
+
+    security_group_ids = get_security_groups(
+        security_group_ids, aws_client('ec2', configuration, secrets))
+
+    client = aws_client('elbv2', configuration, secrets)
+    load_balancers = get_load_balancer_arns(load_balancer_names, client)
+
+    if load_balancers.get('network', []):
+        raise FailedActivity(
+            'Cannot change security groups of network load balancers.')
+
+    results = []
+    for l in load_balancers['application']:
+        response = client.set_security_groups(
+            LoadBalancerArn=l, SecurityGroups=security_group_ids)
+
+        # add load balancer arn to response
+        response['LoadBalancerArn'] = l
+        results.append(response)
+    return results
+
+
+def set_subnets(load_balancer_names: List[str] = None,
+                subnet_ids: List[str] = None,
+                configuration: Configuration = None,
+                secrets: Secrets = None) -> List[AWSResponse]:
+    """
+    Changes the subnets for the specified application load balancer(s)
+    This action will replace the existing security groups on an application
+    load balancer with the specified security groups.
+
+    Parameters:
+        - load_balancer_names: a list of load balancer names
+        - subnet_ids: a list of subnet ids
+
+    returns:
+        [
+            {
+                'LoadBalancerArn': 'string',
+                'AvailabilityZones': {
+                    'ZoneName': 'string',
+                    'SubnetId': 'string',
+                    'LoadBalancerAddresses': [
+                        {
+                            'IpAddress': 'string',
+                            'AllocationId': 'string'
+                        }
+                    ]
+                }
+            },
+            ...
+        ]
+    """
+    if not load_balancer_names:
+        raise FailedActivity('You must specify at least 1 load balancer.')
+
+    if not subnet_ids:
+        raise FailedActivity('You must specify at least 1 subnet id')
+
+    subnet_ids = get_subnets(
+        subnet_ids, aws_client('ec2', configuration, secrets))
+
+    client = aws_client('elbv2', configuration, secrets)
+    load_balancers = get_load_balancer_arns(load_balancer_names, client)
+
+    if load_balancers.get('network', []):
+        raise FailedActivity(
+            'Cannot change subnets of network load balancers.')
+
+    results = []
+    for l in load_balancers['application']:
+        response = client.set_subnets(
+            LoadBalancerArn=l, Subnets=subnet_ids)
+        response['LoadBalancerArn'] = l
+        results.append(response)
+    return results
+
+
 ###############################################################################
 # Private functions
 ###############################################################################
+def get_load_balancer_arns(load_balancer_names: List[str],
+                           client: boto3.client) -> Dict[str, List[str]]:
+    """
+    Returns load balancer arns categorized by the type of load balancer
+
+    return structure:
+    {
+        'network': ['load balancer arn'],
+        'application': ['load balancer arn']
+    }
+    """
+    results = {}
+    logger.debug('Searching for load balancer name(s): {}.'.format(
+        load_balancer_names))
+
+    try:
+        response = client.describe_load_balancers(
+            Names=load_balancer_names)
+
+        for lb in response['LoadBalancers']:
+            if lb['State']['Code'] != 'active':
+                raise FailedActivity(
+                    'Invalid state for load balancer {}: '
+                    '{} is not active'.format(
+                        lb['LoadBalancerName'], lb['State']['Code']))
+            results.setdefault(lb['Type'], []).append(
+                lb['LoadBalancerArn'])
+            results.setdefault('Names', []).append(lb['LoadBalancerName'])
+    except ClientError as e:
+        raise FailedActivity(e.response['Error']['Message'])
+
+    missing_lbs = [l for l in load_balancer_names if l not in results['Names']]
+    if missing_lbs:
+        raise FailedActivity(
+            'Unable to locate load balancer(s): {}'.format(missing_lbs))
+
+    if not results:
+        raise FailedActivity(
+            'Unable to find any load balancer(s) matching name(s): {}'.format(
+                load_balancer_names))
+
+    return results
 
 
 def get_target_group_arns(tg_names: List[str],
@@ -87,3 +235,32 @@ def get_targets_health_description(tg_arns: Dict,
                  .format(str(tg_health_descr)))
 
     return tg_health_descr
+
+
+def get_security_groups(sg_ids: List[str], client: boto3.client) -> List[str]:
+    try:
+        response = client.describe_security_groups(
+            GroupIds=sg_ids)['SecurityGroups']
+        results = [r['GroupId'] for r in response]
+    except ClientError as e:
+        raise FailedActivity(e.response['Error']['Message'])
+
+    missing_sgs = [s for s in sg_ids if s not in results]
+    if missing_sgs:
+        raise FailedActivity('Invalid security group id(s): {}'.format(
+            missing_sgs))
+    return results
+
+
+def get_subnets(subnet_ids: List[str], client: boto3.client) -> List[str]:
+    try:
+        response = client.describe_subnets(SubnetIds=subnet_ids)['Subnets']
+        results = [r['SubnetId'] for r in response]
+    except ClientError as e:
+        raise FailedActivity(e.response['Error']['Message'])
+
+    missing_subnets = [s for s in subnet_ids if s not in results]
+    if missing_subnets:
+        raise FailedActivity('Invalid subnet id(s): {}'.format(
+            missing_subnets))
+    return results